Correcoes

+++ cuda/host.cu
@@ -47,7 +47,7 @@ __global__ void box_filter_kernel_8u_c1(unsigned char* output,const int width, c
             for(int c2=-2; c2<=2; c2++)
             {
             if(l2 >= 0 && c2 >= 0) {
-                     atomicAdd(sum, tex2D(tex8u,inicio+ xIndex+l2,yIndex + c2));
+                     atomicAdd(sum, (float)tex2D(tex8u,inicio+ xIndex+l2,yIndex + c2));
                 //sum += tex2D(tex8u,inicio+ xIndex+l2,yIndex + c2);
                 cont++;
             }
